requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# Data Processing  
pandas==2.2.2
numpy==1.26.4

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0

# API
pydantic==2.5.0
python-multipart==0.0.6

# Database
# sqlite3 is built into Python. No need to install.

# Visualization
plotly==5.17.0
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
# Optional (LLM API, DB)
groq
pymongo

